requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
idna==2.10
python-telegram-bot==13.6
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.5
